There is a demo of running Secretless + pg in Kubernetes (#123)
This PR adds a Secretless in Kubernetes demo to `demos/k8s-demo`. The README includes all needed instructions for running the demo, and gives the option of the end user using their own sample app or an external DB. The demo does include its own sample app and spins up pg in Kubernetes if needed.