Update pg benchmarks to allow for comparison
Update the pg benchmark test so that you can run it against Secretless or the
target service directly. This facilitates comparison with benchcmp. Also update
the test to use IPs instead of container names, to reduce differences in
performance caused by the extra DNS lookup in the Secretless connection. This
includes updating the secretless.yml to get the pg address from the environment.

Also update the test script to allow a `-b` flag to run the benchmark test. This
will enable adding a `test_benchmarks` script to run the benchmarks in each
test directory.