Update the Kubernetes tutorial to use default SSL
Add certificates provided via Kubernetes Secrets to the PostgreSQL server
deployed in the Kubernetes tutorial. At current to add the certificates and
make them readable by the default postgres user/group we must specify the file
permissions of the secrets as 0600 (or as `defaultMode: 384`, which is
equivalent) and specify that the pod runs with the setting
```
securityContext:
  fsGroup: 999
```
At some point if kubernetes fixes the permissions issues referenced in this
GH issue: https://github.com/kubernetes/kubernetes/issues/2630, a different
configuration may be preferable.