Remove debug from all demos and examples
Also updates the secretless main.go to note that the debug flag is for
development purposes only. This caveat should remain in place until such time
as we are auto-reviewing logs to ensure no leakage of secret data even in debug
mode, and the debug values are actually set up to be helpful to an end user who
is trying to debug the broker (as opposed to being useful to a developer, who
has a deep knowledge of the product and is trying to understand its internals).