Updating website build to gen godocs in go img
Previously we were trying to install go-tools to be able to run the godoc
command in the website Docker image, but this was broken when godoc was merged
into the main go binary. To address this, we no longer install go-tools in the
ruby image, but instead build the godocs using the same go image we use for
the project build.