Removing deprecated full-demo
The demos/full-demo has not been validated as a while, and as a project we've
been moving toward cleaner, more lightweight demos. To ensure that the demos
we provide are maintained and relevant for our users, this demo is being removed
in favor of focusing on our other demos in the demos/ folder.