Converted perf agent test runner logic with streaming queue
With this change and some additional architectural changes we are able
to queue up tests and exit out of the streaming properly, allowing us a
much higher degree of parallelism in the code. Just like the earlier
commits, this change sets us up for easier ability to do parallel load
testing of sockets.