Added perf agent ability to exceed round max if not all backends reached it
This change allows us to properly compare stats on multi-threaded
comparisons since it will run parallel runs for the full length of the
time-based or round-based exit triggers. By ensuring that the round max
is reached on all backends we can ensure that we simulated exact load
scenarios on all of them during the full run.