Refactored the code to be able to test the CLI param parsing easier
Old code could not be tested at all since it was living in the main
package so this change splits things up into pieces that are more
suitable for tests.

Note: We may need to turn plugin.Manager into an interface if we want
more testing to be done as we need a "NOOP" manager to avoid running the
real code.