Make StopOnExitSignal synchronous
This will allow clients to decide if they want it to run async or not.
Adding async is as simple as wrapping the call in a go routine.  Whereas
it's useful to be able to block if we want to do, as we do in the main
entrypoint method where Secretless needs to block after starting up all
its services.