Fixed a fatal crash when external plugin folder did not exist
We assumed that the external plugin folder was always present on the
target system but that might not be the case which is a warning and not
a fatal exit condition. We now just show a warning and return no
external plugins if we are running in such an environment.