Support creating an HTTP ProxyService in top-level proxyServices
Before this commit, we could only go end to end with TCP ProxyServices,
because the top-level code that reads the config and creates the
ProxyServices requested in it only supported that.  This adds support
for creating an HTTP ProxyService.

That actual HTTP ProxyService still needs to be created.  But once it
is, this commit means that it will just work.