Reorganize tons of stuff to remove circular dep
Moving the HTTPServiceConfig to the v2.config package, where it
belonged, created _lots_ of problems, which required fix after fix until
all the circular dependencies were removed.

Why?

Fundamentally, separating TCP and HTTP plugins from one another requires
AllAvailalblePlugins(), because the type of an external plugin can only be
determined at runtime (and that logic lives in AllAvailalblePlugins().
So there's no way to logically avoid this.  Though it's probably the
case that we can do some additional re-organization to make things even
better and clearer.

A really good strategy for avoiding these kind of circular dependencies
in the future is described here:

https://youtu.be/MzTcsI6tn-0?t=987