Change Connector API to a proper interface
Prior to this PR, both the tcp and http Connector interfaces were
defined as naked functions.

While this choice was well-intentioned ("nothing is simpler than a
function"), it turned out to create some awkwardness in the code itself,
because we needed to refer to the "connector" (a noun) as well as call
it, which meant doing semantically incorrect things like
x.Connector(request, creds) which doesn't make any literal sense.

A "connector" is still fundamentally a simple function -- that hasn't
changed -- but it is now defined technically as an interface with that
single function on it. This resolves the problems mentioned above and
makes talking about connectors easier and more accurate.