Fix available_plugins test logic and improve mocks
This gives our mock objects and id, so that they can be compared for
identity without worrying about pointers.

It also improves the tests for the override logic, but testing that the
entire result is correct, and not only that the override happened.

Finally, it improves and simplifies the plugin "mock" package.