Simplify signal API and add unit tests for signal package
Srdjan and I both wrote some earlier drafts of unit tests for the signal
package (see this thread for details:
https://conjurhq.slack.com/archives/CBTETAE20/p1572470326051400).

After a few iterations, we realized the complexity of the tests were
arising from the unidiomatic `Wait()` method on the ExitListener API.

We updated that API to follow the convention of returning a "done
channel" instead, allowing client code to block on that.

This simplified the UTs, which became what you see in this commit.