Move socket cleanup into ConfigEnv
We had already created the ConfigEnv to house logic for v2.Config
validation that was about more than simple yml or config structure
validation.

Previously, we'd moved the validation that the requested connectors did
in fact exist in the AvailablePlugins.

In this commit, we also add the logic for deleting old sockets to
ConfigEnv.  We also rename the main method of ConfigEnv to "Prepare".

The idea is that the runtime environment is "preparing itself" (or
erroring) based on the details of the Services requested by in the
Config.