Move ConfigEnv into own file.  Convert to NetworkAddress at border.
The ConfigEnv logic has gotten long enough to bloat up the service file.
It's a self-contained responsibility, and belongs in its own file.

Separately, this commit does a bit of refactoring:

The Service.ListenOn field had been a raw string, which was forcing its
clients to perform low-level string manipulation.  We change its type
from string to NetworkAddress. Clients of the Config now deal with the
NetworkAddress type, which simplifies their code.