Implement generic_http connector, including unit tests
This commit implements the generic_http connector, including basic white
box unit tests for "config", the internal type doing the heavy lifting
of the implementation.

The "config" holds the template parsing and validation logic, and so we
test it directly.  Currently the tests validate the header template
logic, but we should also add a few unit tests for the
credentialPatterns logic.

Considerations:

We might also unit test the connector itself, but since all the
difficult logic is in the config type, this seems unnecessary, given
that we have e2e tests as well.  It felt cleaner to unit test config
directly.