Update dependency checker script
The script was a little confusing when dependencies needed to be updated,
because it wasn't clear how exactly to update the assets/license_finder.txt file

With these changes the file is updated directly and the `git diff` is used to
determine if there is a difference

The CONTRIBUTING directions for updating the license info have also been updated
to be clearer. Even so, it is clear that further improvements can be made to
make this process even more seamless (and in particular, automatable).