Make generic HTTP connector logic usable by other connectors
The generic HTTP connector was already finished prior to this commit,
but it was only available for use by end-users via secretless.yml.

This commit refactors the generic connector so that's its internal logic
is available for other internal connectors to use.  That way, those
other connectors can have a purely declarative, almost trivial
implementation.

Specifically, we do the following:

- Rename the package from "generic" to "generichttp" so its purpose will
  be clear when it's imported.
- Expose "ConfigYAML" so other connectors can write "yaml in code form."
- Expose a meta constructor "NewNewConnector" that allows other
  connectors to generate their own "NewConnector" methods declaratively.
- Adjust the internals of "generichttp" as needed in response to the
  above changes.