Updates Ruby version in docs container
This change addresses a version mismatch between Ruby and
jekyll-sass-converter in the docs container. With the current
implementation, when the docs container is started with:
```
cd docs
docker-compose up -d
```
then an error is reported when the container is being built:
```
ERROR:  Error installing jekyll:
	There are no versions of jekyll-sass-converter (~> 2.0) compatible with your Ruby & RubyGems. Maybe try installing an older version of the gem you're looking for?
	jekyll-sass-converter requires Ruby version >= 2.4.0. The current ruby version is 2.3.7.456.
```

The fix is to upgrade the container to Ruby 2.4.0.

Fixes Issue #1028