Integration tests for mssql have been rewritten to use sqlcmd
This change modifies the integration tests for the mssql connector
such that the sqlcmd is used to test access to an MSSQL server
both directly (to establish a testing baseline) and indirectly
via secretless broker.

The test detects when the `secretless-dev` service is available, and if
so, uses that for testing (instead of `secretless`).

The test cases include:
- Direct access to the MSSQL server with correct credentials succeeds
- Direct access to the MSSQL server with bad credentials fails
- Indirect access to MSSQL server via secretless with bad credentials
  succeeds.

This change fixes secretless-broker Issue #1039.