Add UT scaffolding and happy path UT
This commit sets up the infrastructure for unit testing the mssql
connector and adds the first unit test for the happy path.  This
infrastructure include:

- Type definitions.  These provide abstractions to capture the parts of
  our third-party driver dependency that we care about.
- Mocks.  These provide doubles for those third-party abstractions,
  allowing us to conveniently control them in our tests.

The Secretless MSSQL connector now has no direct dependency on the mssql
3rd party driver: it depends only on our new abstractions.

Some changes to the driver itself were also needed to make this work.