Remove the ability to override an internal plugin with external one
Currently, an external http or tcp connector plugin is allowed to replace
an internal connector plugin of the same type (http/tcp) that uses the
same plugin ID.

This change removes this capability, and forces all external plugins to
use plugin IDs that are unique among all other internal/external plugins.
If there is a name conflict between an external plugin and any other
plugin, then a panic occurs.

Note that there is no need to check for duplicate IDs among `internal`
plugins at run time because these plugins are hardcoded, so that the
only way that duplicate IDs could occur is if the same name
was used twice erroneously.

Addresses Issue #1085