This change adds remote debug capability for the secretless broker. WIth this change, a Delve-capable IDE such as Intellij/Goland can attach to a running secretless broker container, allowing breakpoints to be set, single-step debugging, etc.
The remote debug capability will be helpful for debugging integration tests.
(Unit tests, on the other hand, can already be easily debugged directly from
the source code without remotely connecting to into a container).

This change includes a new "debug" image for the secretless broker that
adds a Delve binary (to allow remote debug connections between the secretless
broker and an IDE), and compiles the secretless broker with code optimizations
disabled (for better debugging experience).

To use remote debugging on the secretless broker, follow instructions
in the 'Debugging Secretless Broker as it is Running in a Container'
section of the test/connector/mssql/README.md file.