Secretless faithfully propagates MSSQL Login Response from server
This change removes any handling and translation of login responses
from the server in the secretless broker. Instead, secretless will
now "call it a day" after sending a translated version of the client's
login request to the MSSQL server. Login responses from the server
will now be passed transparently (from the perspective of secretless)
from the server to the client.