Juxtaposer start script does not error on tmp/juxtaposer_deployment.yml
Currently, if the `bin/juxtaposer/deploy/start` script is run, e.g.
to deploy on OpenShift, the script errors out with a file error trying
to open `deploy/tmp/juxtaposer_deployment.yml`, unless the directory
`deploy/tmp` is manually created beforehand.

This change adds the creation of the `deploy/tmp` dir to the script
before the juxtaposer_deployment.yml is rendered.