Adds SIGPIPE to signals handled by Secretless Juxtaposer
This change adds the SIGPIPE signal to the list of signals that
are handled by the Secretless Juxtaposer test. The intent of the
Juxtaposer signal handler is to catch otherwise fatal signals
so that the Juxtaposer container can stay resident and running
even when fatal errors occur so that logs can be examined.

The SIGPIPE signal is being added to the list of signals being
handled because the following fatal error was observed in a
Juxatposer run after the Secretless sidecar container crashed
(because of authen issues with the Conjur follower), causing the
Juxtaposer container to crash:

```
$ kubectl logs -n dane-secretless-xa juxtaposer-pg-687d4675dc-mld84 juxtaposer-pg
2020/02/04 23:18:35 Juxtaposer starting up...
2020/02/04 23:18:35 Using configuration: /etc/juxtaposer-pg/juxtaposer-pg_pg.yml
2020/02/04 23:18:35 Config loaded!
2020/02/04 23:18:35 Registering shutdown signal listener...
2020/02/04 23:18:35 Using test duration of 96h0m0s (overriding any 'rounds' from configfile)
2020/02/04 23:18:35 Driver: postgres
2020/02/04 23:18:35 Recreate connections: true
2020/02/04 23:18:35 Backends: 3
2020/02/04 23:18:35 Threads: 1
2020/02/04 23:18:35 Rounds: infinity
2020/02/04 23:18:35 Duration: 96h
2020/02/04 23:18:35 Setting up backend: pg_secretless_socket
2020/02/04 23:20:52 ERROR sql: Could not execute query!
2020/02/04 23:20:52 read unix @->/sock/.s.PGSQL.5432: read: connection reset by peer
```

Fixes issue #1131