Use own self-signed certs across MSSQL test environments
Using our own self-signed certs gives us access to the key-pair which is needed for testing. This commit changes the location of the key-pair within the mssql containers to a directory (/certs) that is accessible to the user that the container is being run as, /etc/ssl/private isn't accessible.