test: Add code for mssql in-process proxy service and mock mssql service
Invoking the proxy service, as opposed to the Secretless binary, makes it possible to avoid the perils of static configuration. It gives the integration tester more flexibility in exercising a variety of test cases, all generated in-process.

A mock mssql service makes it possible to capture packets and validate the propagation of parameters.