Update the release instructions for goreleaser
Previously we were manually building the goreleaser artifacts on our local
machines, but in #1234 we started running goreleaser in our build pipeline and
archiving the output. This commit updates the release instructions so that the
releaser will grab the artifacts from the Jenkins output and upload them to the
github release, rather than building them locally.