Update Go version in Dockerfiles to 1.15
We recently updated the Go version in go.mod, but all of our tests run in
Docker and this commit updates the Go version used in our test flows to 1.15 as
well.

Since this upgrades our Debian version to Buster, we also needed to update
the method we use to install openjdk in our MSSQL and PG tests.