Update Go version in Dockerfiles to 1.15
We recently updated the Go version in go.mod, but all of our tests run in
Docker and this commit updates the Go version used in our test flows to 1.15 as
well.

Since this upgrades our Debian version to Buster, we also needed to update
the method we use to install openjdk in our MSSQL and PG tests.

We also had to update our certificates in the MSSQL tests - I changed the
host name in the certs to make clear that it's intentionally set to not match
the host, so that the sslhost tests can verify its ability to override the
cert host.