Change MSSQL repo from Ubuntu to Debian and retry on repo list download failure
Change MSSQL repo from Ubuntu 16.04 to Debian 10 (Buster), which is more appropriate for the Buster container in which the integration test is run.
Retry the repo list download on error - a maximum of 5 times, 10 second timeout.
HTTP error code will now propagate up to curl return code so CI will stop on error