Update dev environment to work on CyberArk NG laptops
On CyberArk dev laptops, golang module dependencies are downloaded with a
corporate proxy in the middle. For these connections to succeed we need to
configure the proxy CA certificate in build containers.

To allow build script to also work on non-CyberArk laptops where the CA
certificate is not available, we copy the (potentially empty) directory
into the build containers, and update container certificates based on that
directory, rather than relying on the CA file itself.